refactor: users module refactored
